requirements.txt
Flask
python-dotenv
cryptography
